requirements.txt
torch
transformers
gradio
Pillow
sentencepiece
huggingface_hub[hf_xet]
